Title: 사각형 채우기 3, Time: 8ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240724/사각형 채우기 3/rectangle-fill-3.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 using namespace std;
 
-int n,dp[1010];
+long long n,dp[1010];
 
 int main() {
